--- src/Templates/Boilerplate/Bit.Boilerplate/src/Client/Boilerplate.Client.Core/Components/Layout/HeaderButtons.razor
@@ -1,28 +1,0 @@
-﻿@inherits AppComponentBase
-
-<BitButton IconOnly
-           FixedColor
-           Variant="BitVariant.Text"
-           OnClick="WrapHandled(ToggleTheme)"
-           Color="BitColor.SecondaryBackground"
-           IconName="@(currentTheme == AppThemeType.Light ? BitIconName.ClearNight : BitIconName.Sunny)" />
-
-@if (CultureInfoManager.MultilingualEnabled)
-{
-    <BitDropdown Items="cultures"
-                 FitWidth NoBorder Transparent
-                 DefaultValue="@CultureInfo.CurrentUICulture.Name"
-                 OnChange="WrapHandled((string c) => OnCultureChanged(c))">
-        <TextTemplate Context="item">
-            <BitImage Src="@($"_content/Boilerplate.Client.Core/images/flags/{item.Value}.webp")" />
-        </TextTemplate>
-        <ItemTemplate Context="item">
-            <BitStack Horizontal VerticalAlign="BitAlignment.Center">
-                <BitImage Src="@($"_content/Boilerplate.Client.Core/images/flags/{item.Value}.webp")" />
-                <BitText Typography="BitTypography.Body1" Style="@(item.IsSelected ? "font-weight:bold" : "")">
-                    @item.Text
-                </BitText>
-            </BitStack>
-        </ItemTemplate>
-    </BitDropdown>
-}
