--- src/Templates/Boilerplate/Bit.Boilerplate/src/Client/Boilerplate.Client.Core/Components/Layout/HeaderButtons.razor.cs
@@ -1,38 +1,0 @@
-﻿namespace Boilerplate.Client.Core.Components.Layout;
-
-public partial class HeaderButtons : AppComponentBase
-{
-    private BitDropdownItem<string>[] cultures = default!;
-
-
-    [AutoInject] private ThemeService themeService = default!;
-    [AutoInject] private CultureService cultureService = default!;
-
-
-    [CascadingParameter] private BitDir? currentDir { get; set; }
-    [CascadingParameter(Name = Parameters.CurrentTheme)] private AppThemeType? currentTheme { get; set; }
-    [CascadingParameter(Name = Parameters.IsCrossLayoutPage)] private bool? isCrossLayoutPage { get; set; }
-
-
-    protected override async Task OnInitAsync()
-    {
-        if (CultureInfoManager.MultilingualEnabled)
-        {
-            cultures = CultureInfoManager.SupportedCultures
-                        .Select(sc => new BitDropdownItem<string> { Value = sc.Culture.Name, Text = sc.DisplayName })
-                        .ToArray();
-        }
-
-        await base.OnInitAsync();
-    }
-
-    private async Task ToggleTheme()
-    {
-        await themeService.ToggleTheme();
-    }
-
-    private async Task OnCultureChanged(string? cultureName)
-    {
-        await cultureService.ChangeCulture(cultureName);
-    }
-}
