+++ src/Templates/Boilerplate/Bit.Boilerplate/src/Client/Boilerplate.Client.Core/Components/Layout/MainLayout.razor.items.cs
@@ -28,13 +28,13 @@ public partial class MainLayout
             {
                 Text = localizer[nameof(AppStrings.SignIn)],
                 IconName = BitIconName.Signin,
-                Url = Urls.SignInPage,
+                Url = $"{Urls.SignUpPage}?return-url={Uri.EscapeDataString(navigationManager.GetRelativePath())}",
             },
             new()
             {
                 Text = localizer[nameof(AppStrings.SignUp)],
                 IconName = BitIconName.Signin,
-                Url = Urls.SignInPage,
+                Url = $"{Urls.SignInPage}?return-url={Uri.EscapeDataString(navigationManager.GetRelativePath())}",
             },
         ];
